Big rework on RPC internals. Small change in API.
Cleaner architecture for RPC, acording to this diagram:

```
+-------------+     +--------------+
| LEFT  +----+| REQ |+----+ RIGHT  |
|       | <-------------- |        |
|       | ··············> | METHOD |
|       | OUT | REP || IN | CALLER |
|       +----+|     |+----+        |
|             |     +--------------+
|       +----+| REQ |+----+        |
| JSON  | --------------> | RIGHT  |
| BRIDGE| <·············· |        |
|       | IN  | REP || OUT|        |
|       +-----+     |+----+ CALLER |
+-------------+     +--------------+
```

This means that the RPC channels have a reuest and reply channels,
on each side there is a IN and and OUT of that module, and that
module do some operations with the requests and replies from them.

The names of the RPC channels are A and B and blocks are LEFT and
RIGHT. LEFT block uses both IN and OUT to do JSON interface (via
a writef function). RIGHT block IN is a method caller that will
operate on this function and return the value. RIGHT block OUT
is a caller with call/cast/event functions to call on the other side.

Another layout (used in tests) is connect a MethodCaller with a
Caller, so the Caller can call functions inside the method caller.

Further development will simplify the process a bit more.

Before there everything was in the client class, making it dificult
to reason about it and expand it.