Reworked method caller, now no recusrive, call first
Now it is based on calls instead of casts+continuation.

Cast+continuation where difficult to reason about, whereas calls are
quite easy. The number of executed processes is the same, and the code
has been split into several functions to actually be able to support
both use cases. Internally it uses calls whenever it is possible.