Use RPC.Client as context holder, writef is MFA
Writef as a {mod, fun, args} allows to simplify some loic to pass the 
client and
the line as arguments. It will receive the original args, then append 
the client
and the line.

This means cleaner definition of the clients, and with client as a 
simple
context holder, there is no need in many cases of special context 
passing.